+++ src/components/SortIndicator/index.ts
@@ -1,7 +1,8 @@
-// SortIndicator from copy-ui @ 2025-10-12.
+// SortIndicator from copy-ui @ 2025-11-07.
 
 export type {
   SortIndicatorDirection,
   SortIndicatorProps,
-} from "./SortIndicator";
-export { SortIndicator, useSortIndicatorState } from "./SortIndicator";
+} from "./sort-indicator";
+export { SortIndicator } from "./sort-indicator";
+export { useDirection } from "./use-direction";