+++ src/components/SortIndicator/CHANGELOG.md
@@ -1,3 +1,5 @@
 - 2025-09-04: Initial version.
 - 2025-09-05: Add the `size` prop.
 - 2025-10-12: Code formatting improvement.
+- 2025-11-07: Support the new hook `useDirection` and remove the old hook. Let
+  the `SortIndicator` has its style.